requirements.txt
pyautogen>=0.2.3
python-dotenv>=1.0.0
anyio>=4.0.0
asyncio>=3.4.3
nest-asyncio>=1.5.8
pydantic>=2.0.0
mcp 